Refactored cache
