+++ app/cache/get-request-index.js
@@ -1,0 +1,7 @@
+const { isDeepStrictEqual } = require('util')
+
+const getRequestIndex = (cacheData, request) => {
+  return cacheData.requests.findIndex(x => isDeepStrictEqual(x.request, request))
+}
+
+module.exports = getRequestIndex
