+++ app/cache/set-cached-response.js
@@ -1,0 +1,11 @@
+const { get, update } = require('./base')
+const getRequestIndex = require('./get-request-index')
+
+const setCachedResponse = async (cacheName, key, request, response) => {
+  const cacheData = await get(cacheName, key)
+  const requestIndex = getRequestIndex(cacheData, request)
+  cacheData.requests[requestIndex].response = response
+  await update(cacheName, key, cacheData)
+}
+
+module.exports = setCachedResponse
