+++ app/messaging/process-calculate-message.js
@@ -1,4 +1,4 @@
-const cache = require('../cache')
+const { getCachedResponse, setCachedResponse } = require('../cache')
 const { calculatePaymentRates } = require('../calculate')
 const config = require('../config')
 const sendMessage = require('./send-message')
@@ -8,14 +8,14 @@ const processCalculateMessage = async (message, receiver) => {
     const { body, correlationId, messageId } = message
     const { code, parcels } = body
 
-    const cachedResponse = await cache.getCachedResponse('calculate', body, correlationId)
+    const cachedResponse = await getCachedResponse('calculate', body, correlationId)
 
     // if request already processed then return without reprocessing
     if (cachedResponse) {
       await sendMessage(cachedResponse, 'uk.gov.sfi.agreement.calculate.response', undefined, messageId, config.calculateResponseQueue)
     } else {
       const paymentRates = calculatePaymentRates(code, parcels)
-      await cache.updateCache('calculate', correlationId, body, paymentRates)
+      await setCachedResponse('calculate', correlationId, body, paymentRates)
       await sendMessage(paymentRates, 'uk.gov.sfi.agreement.calculate.response', undefined, messageId, config.calculateResponseQueue)
     }
     await receiver.completeMessage(message)
